src/apiproxy/pyproject.toml
[tool.uv.sources]
apiproxy = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["."]

[project]
name = "ApiProxy"
version = "1.1.0"
description = "A Python package with a built-in web application"
requires-python = ">=3.10,<3.13"
license = "MIT"
keywords = ["ApiProxy", "OpenAI"]
readme = "README.md"
maintainers = [
    { name = "Neeker", email = "13317312768@qq.com" },
]

# Define your main dependencies here
dependencies = [
    "pyyaml>=6.0.2,<7.0.0",
    "types-pyyaml>=6.0.12.8",
    "numpy>=1.26.4,<2.0.0",
    "fastapi>=0.115.2,<1.0.0",
    "httpx[http2]>=0.27,<1.0.0",
    "uvicorn>=0.30.0,<1.0.0",
    "pydantic-settings>=2.2.0,<3.0.0",
    "pydantic>=2.10.0,<2.11.0",
    "gunicorn>=22.0.0,<24.0.0",
    "opentelemetry-api>=1.25.0,<2.0.0",
    "opentelemetry-sdk>=1.25.0,<2.0.0",
    "opentelemetry-exporter-prometheus>=0.46b0,<1.0.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0,<1.0.0",
    "prometheus-client>=0.20.0,<1.0.0",
    "cachetools>=5.3.0,<6.0.0",
    "types-cachetools>=5.3.0.5,<6.0.0",
    "rich>=13.7.0,<14.0.0",
    "sentry-sdk[fastapi,loguru]>=2.5.1,<3.0.0",
    "passlib>=1.7.4,<2.0.0",
    "bcrypt==4.0.1",
    "shortuuid>=1.0.13,<2.0.0",
    "fire>=0.7.0,<1.0.0",
    "requests>=2.0.0,<3.0.0",
    "orjson==3.10.0",
    "alembic>=1.13.0,<2.0.0",
    "multiprocess>=0.70.14,<1.0.0",
    "cachetools>=5.5.0,<6.0.0",
    "platformdirs>=4.2.0,<5.0.0",
    "filelock>=3.15.4,<4.0.0",
    "sqlmodel==0.0.18",
    "sqlalchemy[aiosqlite,postgresql_psycopg2binary,postgresql_psycopgbinary]>=2.0.36,<3.0.0",
]

[project.urls]
Repository = "https://snz1.cn/gitrepo/dp/ai/apiproxy"
Documentation = "https://docs.dingtalk.com/i/spaces/nb9XJZ4VdyeYVXyA/overview"

[project.scripts]
apiproxy = "openaiproxy.main:__main__"

[tool.uv]
dev-dependencies = [
    "pyyaml>=6.0.2,<7.0.0",
    "types-pyyaml>=6.0.12.8",
    "numpy>=1.26.4,<2.0.0",
    "fastapi>=0.115.2,<1.0.0",
    "httpx[http2]>=0.27,<1.0.0",
    "uvicorn>=0.30.0,<1.0.0",
    "pydantic-settings>=2.2.0,<3.0.0",
    "pydantic>=2.10.0,<2.11.0",
    "gunicorn>=22.0.0,<24.0.0",
    "opentelemetry-api>=1.25.0,<2.0.0",
    "opentelemetry-sdk>=1.25.0,<2.0.0",
    "opentelemetry-exporter-prometheus>=0.46b0,<1.0.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0,<1.0.0",
    "prometheus-client>=0.20.0,<1.0.0",
    "cachetools>=5.3.0,<6.0.0",
    "types-cachetools>=5.3.0.5,<6.0.0",
    "rich>=13.7.0,<14.0.0",
    "sentry-sdk[fastapi,loguru]>=2.5.1,<3.0.0",
    "passlib>=1.7.4,<2.0.0",
    "bcrypt==4.0.1",
    "shortuuid>=1.0.13,<2.0.0",
    "fire>=0.7.0,<1.0.0",
    "requests>=2.0.0,<3.0.0",
]

[tool.codespell]
skip = '.git,*.pdf,*.svg,*.pdf,*.yaml,*.ipynb,poetry.lock,*.min.js,*.css,package-lock.json,*.trig.,**/node_modules/**,./stuff/*,*.csv'
# Ignore latin etc
ignore-regex = '.*(Stati Uniti|Tense=Pres).*'

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests", "integration"]
console_output_style = "progress"
filterwarnings = ["ignore::DeprecationWarning", "ignore::ResourceWarning"]
log_cli = true
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
markers = ["async_test", "api_key_required"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
command_line = """
    -m pytest --ignore=tests/integration
    --cov --cov-report=term --cov-report=html
    --instafail -ra -n auto -m "not api_key_required"
"""
source = ["./"]
omit = ["*/alembic/*", "tests/*", "*/__init__.py"]


[tool.coverage.report]
sort = "Stmts"
skip_empty = true
show_missing = false
ignore_errors = true


[tool.coverage.html]
directory = "coverage"


[tool.ruff]
line-length = 120

[tool.ruff.lint]
pydocstyle.convention = "google"
select = ["ALL"]
ignore = [
    "C90", # McCabe complexity
    "CPY", # Missing copyright
    "COM812", # Messes with the formatter
    "ERA", # Eradicate commented-out code
    "FIX002", # Line contains TODO
    "ISC001", # Messes with the formatter
    "PERF203", # Rarely useful
    "PLR09", # Too many something (arg, statements, etc)
    "RUF012", # Pydantic models are currently not well detected. See https://github.com/astral-sh/ruff/issues/13630
    "TD002", # Missing author in TODO
    "TD003", # Missing issue link in TODO
    "TRY301", # A bit too harsh (Abstract `raise` to an inner function)

    # Rules that are TODOs
    "ANN",
]

# Preview rules that are not yet activated
external = ["RUF027"]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = [
    "D1",
    "INP",
    "T201",
]
"src/backend/tests/*" = [
    "D1",
    "PLR2004",
    "S101",
    "SLF001",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
follow_imports = "skip"
disable_error_code = ["type-var"]
namespace_packages = true
mypy_path = "apiproxy"
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
